requirements.txt
flask_sqlalchemy==3.1.1
flasgger==0.5.2
psycopg2-binary==2.9.9